+++ src/components/MyMatches.vue
@@ -57,7 +57,7 @@
 <script>
 import { mapActions, mapGetters } from 'vuex';
 import BreakpointsCond from '../mixins/BreakpointsCond';
-import MatchCardsGroup from './MatchCardsGroup.vue';
+import MatchCardsGroup from './matchesCards/MatchCardsGroup.vue';
 
 export default {
   name: 'myMatches',
